Move EcoClient to demo-gallery
