+++ Autobahn/build.gradle
@@ -22,4 +22,3 @@ dependencies {
     compile 'org.codehaus.jackson:jackson-core-asl:1.9.7'
     compile 'org.codehaus.jackson:jackson-mapper-asl:1.9.7'
 }
-
