--- Demo/EchoClient/.classpath
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="src" path="gen"/>
-	<classpathentry kind="con" path="com.android.ide.eclipse.adt.ANDROID_FRAMEWORK"/>
-	<classpathentry exported="true" kind="con" path="com.android.ide.eclipse.adt.LIBRARIES"/>
-	<classpathentry kind="lib" path="/Autobahn/libs/jackson-core-asl-1.9.7.jar"/>
-	<classpathentry kind="lib" path="/Autobahn/libs/jackson-mapper-asl-1.9.7.jar"/>
-	<classpathentry kind="output" path="bin/classes"/>
-</classpath>
