--- Demo/EchoClient/.gitignore
@@ -1,3 +1,0 @@
-bin/*
-gen/*
-!gen/PLACEHOLDER
