+++ build.gradle
@@ -12,4 +12,3 @@ allprojects {
         jcenter()
     }
 }
-
