+++ demo-gallery/build.gradle
@@ -1,0 +1,28 @@
+apply plugin: 'com.android.application'
+
+android {
+    compileSdkVersion 25
+    buildToolsVersion "25.0.2"
+
+    packagingOptions {
+        exclude 'META-INF/LICENSE'
+        exclude 'META-INF/ASL2.0'
+    }
+
+    defaultConfig {
+        minSdkVersion 10
+        targetSdkVersion 25
+    }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+    return true
+}
+
+dependencies {
+    compile project(path: ':Autobahn')
+    compile 'com.android.support:appcompat-v7:25.3.1'
+}
