+++ demo-gallery/src/main/res/layout/main.xml
