+++ demo-gallery/src/main/res/menu/main_menu.xml
