+++ demo-gallery/src/main/res/values/colors.xml
@@ -1,5 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+    <color name="colorPrimary">#3F51B5</color>
+    <color name="colorPrimaryDark">#303F9F</color>
+    <color name="colorAccent">#FF4081</color>
     <color name="autobahn">#028ec9</color>
     <color name="lightgrey">#cccccc</color>
     <color name="darkgrey">#333333</color>
