+++ demo-gallery/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<resources>
+    <string name="app_name">Autobahn Demo Gallery</string>
+    <string name="button_text_websocket_gallery">WebSocket Gallery</string>
+    <string name="button_text_wamp_gallery">WAMP Gallery</string>
+</resources>
